requirements.txt
# Core Data Processing
pandas==2.0.3
numpy==1.24.3
scipy==1.11.2

# Machine Learning
scikit-learn==1.3.0
scikit-surprise==1.1.3
# lightfm==1.17   # optional

# Visualization
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.16.1

# Web Framework
streamlit==1.52.1
streamlit-option-menu==0.3.6

# Database
sqlalchemy==2.0.20

# Utilities
python-dotenv==1.0.0
tqdm==4.66.1
joblib==1.3.2
kaggle==1.5.16

# Natural Language Processing
gensim==4.3.1   # main version
# gensim==4.3.0 # optional alternative

# Jupyter Notebook Support
jupyter
notebook
ipykernel
